# lib/api-zod/src/generated/types/sendOpenaiMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI SaaS Toolkit API
 * OpenAPI spec version: 0.1.0
 */

export interface SendOpenaiMessageBody {
  content: string;
  systemPrompt?: string;
}